Adds model translation command
Adds a new command to translate model data between languages.

This command facilitates the translation of database models,
allowing for automated translation of content from a source language
to one or more target languages. It improves multilingual support
and reduces manual translation efforts.

The changes also include a minor update to the `TranslationExistException`
to include the locale in the message, and some refactoring
of the Translatable trait, including renaming `Translater` to `Translator`
and handling translation errors.